requirements.txt
cloudinary
python-dotenv 